refactor: :technologist: change Language Switcher component to Smart and Dumb component concept
